requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
gspread==5.12.4
oauth2client==4.1.3
python-dotenv==1.0.1
pandas
fuzzywuzzy
python-Levenshtein
oauth2client